feature: csv and cypher query generation
Create a .csv file based on the emulated revisions. The .csv format is:
topic,timestamp,[list, of, links, to, the, given, topic]
another_topic,timestamp,[only, four, links, here]

Note that the list length is arbitrary.

Also add a module for creating cypher queries. As of now, it only prints
the results to the screen.