feat: influx multi page query (#8)
Allows querying for several ids. One query fetches all views for the given ids, while the other one fetches the average page view for those pages in the given time interval.

Flux allows for contains() filtering (https://community.grafana.com/t/grafana-influxdb-flux-query-for-displaying-multi-select-variable-inputs/35536/13), but this is not efficient when fields are queried for. Therefore, I introduce some regex logic which was proposed in the Grafana discussion. The regex is on the form ^(x|y|z).